opencv_deployment/requirements.txt
argcomplete==2.0.0
attrs==21.4.0
bcrypt==3.2.2
Beaker==1.10.0
beautifulsoup4==4.11.0
blivet==3.4.3
blivet-gui==2.3.0
Brlapi==0.8.3
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
chrome-gnome-shell==0.0.0
click==8.0.4
cryptography==36.0.0
cupshelpers==1.0
cycler==0.11.0
dasbus==1.6
dbus-python==1.2.18
distro==1.6.0
docker==5.0.2
docker-compose==1.29.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
engineering-notation==0.6.0
fedora-third-party==0.10
fonttools==4.33.3
fros==1.1
gpg==1.15.1
humanize==3.13.1
idna==3.3
jsonschema==3.2.0
kiwisolver==1.4.2
langtable==0.0.58
libcomps==0.1.18
lxml==4.7.1
Mako==1.1.4
MarkupSafe==2.0.0
matplotlib==3.5.2
nftables==0.1
numpy==1.22.4
olefile==0.46
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
paramiko==2.11.0
Paste==3.5.0
pexpect==4.8.0
pid==2.2.3
Pillow==9.1.1
ply==3.11
productmd==1.33
ptyprocess==0.6.0
pwquality==1.4.4
pyasn1==0.4.8
pycairo==1.21.0
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.45.1
pyenchant==3.2.2
PyGObject==3.42.1
pykickstart==3.36
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyparted==3.12.0
pyrsistent==0.18.1
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.1
python-dotenv==0.19.2
python-meh==0.50
pytz==2022.1
pyudev==0.22.0
pyxdg==0.27
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
requests-file==1.5.1
requests-ftp==0.3.1
rpm==4.17.0
scipy==1.8.1
seaborn==0.11.2
selinux==3.3
sepolicy==3.3
setools==4.4.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.3
soupsieve==2.3.1
systemd-python==234
Tempita==0.5.2
texttable==1.6.2
tk==0.1.0
tk-tools==0.15.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.8
websocket-client==1.2.3
youtube-dl==2021.12.17
